refactor(solana): extract event handling
