+++ chain-signatures/chain-solana/src/lib.rs
@@ -1,9 +1,11 @@
 mod client;
 mod config;
+mod events;
 mod indexer;
 mod utils;
 
 pub use anchor_lang::prelude::Pubkey;
 pub use client::SolanaClient;
 pub use config::SolConfig;
+pub use events::{emit_events, extract_tx_signature};
 pub use indexer::SolanaIndexer;
